restrict Renderable interface to avoid cycle dependency
If `Render()` takes a `Profile`, `profile` module should be imported.
This produces a cycle dependency.

As of now, creating a specific interface to pass to `Render()` to avoid
this does not provide much gain, so we restrict the argument to
`Render()` to only profile name and location, the only 2 really used.

I can't foresee now if more would be required, but in that case a
specific interface may be created.